Fixed field values of variations are saved as arrays into parent product.
When saving a variable product, the parent product is saved in the same request as the child product variations in a single POST data array. As the product meta field name is the same for the parent product and the variations, the POST data for the field contains a list of values for the variations (instead of the parent product).

If the meta field will also become required for parent products in the future, the name of the meta field used by variations needs to be renamed to contain 'variation', so that the field values of variations do not overwrite the value of the parent product.